equity_pipeline/requirements.txt
numpy==2.1.3
pandas==2.2.3
yfinance==0.2.54
finnhub-python==2.4.24
fredapi==0.5.2
edgartools==3.11.3
pandas==2.2.3
numpy==2.1.3
PyYAML==6.0.2
python-dotenv==1.0.1
yfinance==0.2.55
finnhub-python==2.4.22
fredapi==0.5.2
edgartools==3.5.0
